refactor: updated schema imports to the root schema and not to the actual schema export
