+++ imports/plugins/core/templates/lib/collections/schemas/emailtemplates.js
@@ -2,7 +2,6 @@ import SimpleSchema from "simpl-schema";
 import { check } from "meteor/check";
 import { Tracker } from "meteor/tracker";
 import { Templates } from "/lib/collections";
-import { shopIdAutoValue } from "/lib/collections/schemas/helpers";
 import { registerSchema } from "@reactioncommerce/schemas";
 
 /**
@@ -15,7 +14,6 @@ export const EmailTemplates = new SimpleSchema({
   "shopId": {
     type: String,
     index: 1,
-    autoValue: shopIdAutoValue,
     label: "Template ShopId"
   },
   "name": {
