+++ imports/plugins/included/search-mongo/lib/collections/schemas/search.js
@@ -1,5 +1,5 @@
 import SimpleSchema from "simpl-schema";
-import { PackageConfig } from "/lib/collections/schemas/registry";
+import { PackageConfig } from "/lib/collections/schemas";
 import { registerSchema } from "@reactioncommerce/schemas";
 
 /**
