+++ imports/plugins/included/shipping-shippo/lib/collections/schemas/shippo.js
@@ -1,4 +1,4 @@
-import { PackageConfig } from "/lib/collections/schemas/registry";
+import { PackageConfig } from "/lib/collections/schemas";
 import { registerSchema } from "@reactioncommerce/schemas";
 
 /**
